File structure reorganization and renamings
