+++ neovmUtils/Classes/BlockchainInterfaceProtocol.swift
@@ -8,7 +8,7 @@
 
 import Foundation
 
-protocol BlockchainInterfaceProtocl {
+protocol BlockchainInterfaceProtocol {
     var testnetExecution: Bool { get set }
     func invoke(contractHash: String, operation: String, args: [Any], wallet: Wallet, other: [String: Any]) -> String
     func read(contractHash: String, operation: String, args: [Any]) -> String
